Linking edit/delete room buttons w/ backend
You can now edit and delete rooms from the UI. The trick is sneaking in the roomID to the backend via the Redux state.room.

There are some rehydrating issues. When a rehydrate happens, sometimes room takes on the values of user or vice versa. This can cause you to be effectively logged out.