Coordinates tracked correctly
Will be putting them in Redux state, TokenCoordinates, which is an object of objects. The inner object is coordinates that are assigned to a key equal to the piece's key for the outer object. When a token loads, it'll check if there's a value for that piece's key, and start with those coordinates if so. This will also load any time a socket comes in (a useEffect with a dependency = TokenCoordinates).

Now I just need to web socket it up.